mgr/cephadm: move OSD related code to OSDService
